refactor, aspect logger and slf4j added, repository caching and scheduled clearing added
