--- src/main/java/pl/kuba565/RepositoryInfo/RepositoryInfoApplication.java
@@ -1,13 +1,0 @@
-package pl.kuba565.RepositoryInfo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RepositoryInfoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(RepositoryInfoApplication.class, args);
-	}
-
-}
