--- src/main/java/pl/kuba565/RepositoryInfo/model/RepositoryInfo.java
@@ -1,18 +1,0 @@
-package pl.kuba565.RepositoryInfo.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RepositoryInfo {
-    private String fullName;
-    private String description;
-    private String cloneUrl;
-    private Integer stars;
-    private String createdAt;
-}
