--- src/main/java/pl/kuba565/RepositoryInfo/model/RepositoryRequest.java
@@ -1,13 +1,0 @@
-package pl.kuba565.RepositoryInfo.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class RepositoryRequest {
-    private String owner;
-    private String repositoryName;
-}
