--- src/main/java/pl/kuba565/RepositoryInfo/swagger/SwaggerConfig.java
@@ -1,17 +1,0 @@
-package pl.kuba565.RepositoryInfo.swagger;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@Configuration
-@EnableSwagger2
-public class SwaggerConfig {
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2);
-    }
-}
