+++ src/main/java/pl/kuba565/repositoryInfo/model/RepositoryInfo.java
@@ -1,0 +1,18 @@
+package pl.kuba565.repositoryInfo.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class RepositoryInfo {
+    private String fullName;
+    private String description;
+    private String cloneUrl;
+    private Integer stars;
+    private String createdAt;
+}
