+++ src/main/java/pl/kuba565/repositoryInfo/util/UrlUtil.java
@@ -1,0 +1,9 @@
+package pl.kuba565.repositoryInfo.util;
+
+import pl.kuba565.repositoryInfo.model.RepositoryRequest;
+
+public class UrlUtil {
+    public static String getUrl(RepositoryRequest repositoryRequest) {
+        return String.format("https://api.github.com/repos/%s/%s", repositoryRequest.getOwner(), repositoryRequest.getRepositoryName());
+    }
+}
