+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+logging.level.root=INFO
+logging.file=Application.log
